odoo-eergy-services/requirements.txt
# DTE Microservice - Python Dependencies
# Python 3.11+

# ═══════════════════════════════════════════════════════════
# WEB FRAMEWORK
# ═══════════════════════════════════════════════════════════
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ═══════════════════════════════════════════════════════════
# RABBITMQ - FASE 2
# ═══════════════════════════════════════════════════════════
aio-pika==9.3.0
tenacity==8.2.3

# ═══════════════════════════════════════════════════════════
# HTTP CLIENT - FASE 2 (para webhook a Odoo)
# ═══════════════════════════════════════════════════════════
httpx>=0.25.2

# ═══════════════════════════════════════════════════════════
# XML PROCESSING (CRÍTICO)
# ═══════════════════════════════════════════════════════════
lxml>=4.9.0
xmlsec>=1.3.13
defusedxml>=0.7.1

# ═══════════════════════════════════════════════════════════
# FIRMA DIGITAL Y CRIPTOGRAFÍA (CRÍTICO)
# ═══════════════════════════════════════════════════════════
pyOpenSSL>=23.3.0
cryptography>=41.0.7
asn1crypto>=1.5.1

# ═══════════════════════════════════════════════════════════
# SOAP Y HTTP (CRÍTICO)
# ═══════════════════════════════════════════════════════════
zeep>=4.2.1
requests>=2.31.0
urllib3>=2.1.0
httpx>=0.25.2

# ═══════════════════════════════════════════════════════════
# ASYNC QUEUE
# ═══════════════════════════════════════════════════════════
celery>=5.3.4
redis>=5.0.1
kombu>=5.3.4

# ═══════════════════════════════════════════════════════════
# LOGGING Y MONITOREO
# ═══════════════════════════════════════════════════════════
structlog>=23.2.0
prometheus-client>=0.19.0

# ═══════════════════════════════════════════════════════════
# RATE LIMITING (FIX A3)
# ═══════════════════════════════════════════════════════════
slowapi>=0.1.9

# ═══════════════════════════════════════════════════════════
# UTILIDADES
# ═══════════════════════════════════════════════════════════
python-dotenv>=1.0.0
python-dateutil>=2.8.2
pytz>=2023.3

# ═══════════════════════════════════════════════════════════
# RETRY LOGIC
# ═══════════════════════════════════════════════════════════
tenacity>=8.2.3

# ═══════════════════════════════════════════════════════════
# SCHEDULED TASKS (Polling Automático + Disaster Recovery)
# ═══════════════════════════════════════════════════════════
apscheduler>=3.10.4

# ═══════════════════════════════════════════════════════════
# CLOUD STORAGE (Disaster Recovery - S3 Backups, opcional)
# ═══════════════════════════════════════════════════════════
boto3>=1.34.0

# ═══════════════════════════════════════════════════════════
# AUTHENTICATION & SECURITY (Sprint 1)
# ═══════════════════════════════════════════════════════════
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# ═══════════════════════════════════════════════════════════
# TESTING
# ═══════════════════════════════════════════════════════════
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.25.2

python-stdnum==1.19       # Validación RUT (delegación a Odoo nativo)
